Moves MonetaryAmountProducer and all implementations to function package
