+++ src/test/java/org/javamoney/moneta/function/MoneyProducerTest.java
@@ -1,4 +1,4 @@
-package org.javamoney.moneta.spi;
+package org.javamoney.moneta.function;
 
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertTrue;
@@ -9,11 +9,13 @@ import javax.money.CurrencyUnit;
 import javax.money.Monetary;
 import javax.money.MonetaryAmount;
 
-import org.javamoney.moneta.FastMoney;
+import org.javamoney.moneta.Money;
+import org.javamoney.moneta.function.MonetaryAmountProducer;
+import org.javamoney.moneta.function.MoneyProducer;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
-public class FastMoneyProducerTest {
+public class MoneyProducerTest {
 
 	private MonetaryAmountProducer producer;
 
@@ -21,7 +23,7 @@ public class FastMoneyProducerTest {
 
 	@BeforeMethod
 	public void setup() {
-		producer = new FastMoneyProducer();
+		producer = new MoneyProducer();
 		currency = Monetary.getCurrency(Locale.getDefault());
 	}
 
@@ -44,10 +46,10 @@ public class FastMoneyProducerTest {
 	}
 
 	@Test
-	public void shouldCreateUsingFastMoneyImplementation() {
+	public void shouldCreateUsingMoneyImplementation() {
 		Long value = 10L;
 		MonetaryAmount amount = producer.create(currency, value);
-		assertTrue(FastMoney.class.isInstance(amount));
+		assertTrue(Money.class.isInstance(amount));
 	}
 
 }
